fix: Resolve nil pointer panic, missing error propagation, and ambiguous SQL operator
- Handle nil document in SDA handleGet/handleHead to return 404 instead of panicking
- Return error from client PutDocument on 4xx/5xx server responses (fixes conditional put)
- Cast bitwise NOT parameter to integer in UpdateMessageFlags to resolve PostgreSQL operator ambiguity

+++ internal/storage/postgres/postgres.go
@@ -295,7 +295,7 @@ func (s *PostgresStorage) GetMessageCount(ctx context.Context, mailboxID int64) 
 func (s *PostgresStorage) UpdateMessageFlags(ctx context.Context, messageID string, addFlags, removeFlags uint32) (bool, error) {
 	tag, err := s.pool.Exec(ctx, `
 		UPDATE stored_messages
-		SET flags_bitmap = (flags_bitmap | $1) & ~$2
+		SET flags_bitmap = (flags_bitmap | $1) & ~$2::integer
 		WHERE message_id = $3`,
 		addFlags, removeFlags, messageID,
 	)
